requirements.txt
# CHARIS AI Backend — Python Dependencies
# FastAPI + ASGI server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Data validation & serialization
pydantic>=2.7.0

# Async HTTP client (Anthropic / Stability AI API 호출)
httpx>=0.27.0

# Anthropic SDK (Claude 3.7 Sonnet)
anthropic>=0.28.0

# Image processing (Vision 전처리)
Pillow>=10.3.0

# SQLite ORM (MVP) — async 지원
sqlmodel>=0.0.19

# Environment variable management
python-dotenv>=1.0.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
